GanjeIsar/requirement.txt
#django
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

#database

#img
pillow==11.0.0

#sms
smsir-python==1.0.8